+++ src/datahub_tools/classes.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import logging
 import subprocess
+from collections import defaultdict
 from dataclasses import dataclass
 from typing import Any, Dict, List
 
@@ -103,11 +104,11 @@ class DHEntity(DH):
     def from_dict(cls, _dict: Dict[str, Any]):
         # used to parse the results from query to the graphql search endpoint
         raw_owners = jmespath.search("ownership.owners", _dict) or []
-        owner_urns_by_type = {}
+        owner_urns_by_type = defaultdict(list)
 
         for raw_owner in raw_owners:
             owner_type = raw_owner["type"]
-            owner_urns = owner_urns_by_type.get(owner_type, [])
+            owner_urns = owner_urns_by_type[owner_type]
             owner_urns.append(raw_owner["owner"]["urn"])
             if owner_type not in owner_urns_by_type:
                 owner_urns_by_type[owner_type] = owner_urns